DONE -- CRUD Supplier/Color/ --> TODO need to done CRUD Role / Config output format

+++ src/main/java/com/example/managementapi/Service/SupplierService.java
@@ -1,11 +1,12 @@
 package com.example.managementapi.Service;
 
 
-import com.example.managementapi.Dto.Request.CreateSupplierReq;
-import com.example.managementapi.Dto.Request.UpdateSupplierReq;
-import com.example.managementapi.Dto.Response.CreateSupplierRes;
-import com.example.managementapi.Dto.Response.GetSupplierRes;
-import com.example.managementapi.Dto.Response.UpdateSupplierRes;
+import com.example.managementapi.Dto.Request.Supplier.CreateSupplierReq;
+import com.example.managementapi.Dto.Request.Supplier.UpdateSupplierReq;
+import com.example.managementapi.Dto.Response.Supplier.CreateSupplierRes;
+import com.example.managementapi.Dto.Response.Supplier.GetSupplierDetailRes;
+import com.example.managementapi.Dto.Response.Supplier.GetSupplierRes;
+import com.example.managementapi.Dto.Response.Supplier.UpdateSupplierRes;
 import com.example.managementapi.Entity.Supplier;
 import com.example.managementapi.Mapper.SupplierMapper;
 import com.example.managementapi.Repository.SupplierRepository;
@@ -32,6 +33,12 @@ public class SupplierService {
                 .map(supplier -> supplierMapper.toGetSuppliers(supplier)).toList();
     }
 
+    public GetSupplierDetailRes getSupplierDetailRes(String supplierId){
+        return supplierMapper.toGetSupplierDetailRes(supplierRepository
+                .findById(supplierId)
+                .orElseThrow(() -> new RuntimeException("Supplier not found")));
+    }
+
     public CreateSupplierRes createSupplier(CreateSupplierReq request){
 
        var supplier = supplierMapper.toCreateSupplierReq(request);
@@ -41,9 +48,12 @@ public class SupplierService {
         return supplierMapper.toCreateSupplierRes(supplier);
     }
 
-    public UpdateSupplierRes updateSupplier(String supplier_id, UpdateSupplierReq request){
+    public UpdateSupplierRes updateSupplier(String supplierId, UpdateSupplierReq request){
 
-        Supplier supplier = supplierRepository.findById(supplier_id).orElseThrow(()
+        log.warn("supplier_id_service: " + supplierId);
+
+
+        Supplier supplier = supplierRepository.findById(supplierId).orElseThrow(()
                 -> new RuntimeException("Supplier not found"));
 
         supplierMapper.toUpdateSupplierReq(supplier, request);
@@ -53,6 +63,11 @@ public class SupplierService {
     }
 
     public void deleteSupplier(String supplier_id){
+
+        if(!supplierRepository.existsById(supplier_id)){
+            throw new RuntimeException("Supplier not found");
+        }
+
         supplierRepository.deleteById(supplier_id);
     }
